WIP: Permit refancy of raw to deferred_ptrs
== BEGIN FAIL ==

Disregard this change. I updated my test and realized this
solution does currently does not work.

Work In Progress.

WARNING: Only *my* use cases have been evaluated (`deferred_vector` on
VS2019). This is likely not a "good" change for you.

== END FAIL ==

In VS2019 (and VS2017) the existing code would fail to compile. The
MSVC standard library implementation `_Unwrap()`s iterators to their raw
types, if it can. It subsequently `_Unfancy()`s the contained type. This
in effect bypasses the `deferred_ptr` directly to it's raw counterpart.

The problem arises when an algorithm is finished and wishes to
`_Refancy()` the raw pointer. This invokes `pointer_to()` which has no
contextual information about which heap (if any) this address belongs.

This issue did not present in VS2015 as the library implementation had
not yet utilized `pointer_to` and had instead remained using iterator
types.

A solution is attempted where instead of `pointer_to()` returning a
`deferred_ptr<T>`, it instead returns a `refancy_proxy`. This object
leverages the fact that the MSVC standard library attempts to
`_Seek_to()` the address from the original iterator. The
copy-constructor and assignment operator are abused with this interim
type until such a time that a heap is known.

When an iterator is being updated with a `refancy_proxy`, it first
checks its heap to ensure the desired pointer is owned by the heap.

The solution presented here may not be ideal but it works for now with
minimal changes. It may be philosophically inconsistent with the rest of
the code base. Very limited testing has been done and this design may
have poor performance ramifications.